Resumes logic iniciada

+++ src/main/java/com/softeams/poSystem/core/dtos/ResumeDashboardDto.java
@@ -1,11 +1,12 @@
 package com.softeams.poSystem.core.dtos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ResumeDashboardDto(
         Integer totalProductos,
         Integer totalVentas,
-        Integer ingresosTotales,
+        BigDecimal ingresosTotales,
         Integer stockBajo,
         List<SaleResponse> ventasRecientes
 ) {
